folders moved around and requirements updated
